Made it so that only "AssignCruiseToPassangerMenu" or "ViewPassangerCruiseMenuItem" only show in the ViewPassangerMenu to prevent the customer getting charged double from being assigned the same or different cruises while already being assigned one.
Changed how the CostOfTrip is calculated. Instead of being added and substracted when the Cruise is Assigned/unAssigned it is now done with a method "CalculateCostOfTrip" instead.

Added "TripsThatDontComeFree" list to pasanger so that i can differentiate which trips' cost should be added on when calculating CostOfTrip

Added TripCost attribute to the Trip class and edited the AddTrip MenuItem so that it also asks for the TripCost

Changed the types of all "Cost" type attributes from double to decimal.

Rounded the CostOfTrip to 2 decimal places when displaying the CostOfTrip